+++ app/src/main/java/iti/android/wheatherappjetpackcompose/dataLayer/source/dto/Daily.kt
@@ -3,7 +3,7 @@ package iti.android.wheatherappjetpackcompose.dataLayer.source.dto
 data class Daily(
     val clouds: Double,
     val dew_point: Double,
-    val dt: Long,
+    val dt: Int,
     val feels_like: FeelsLike,
     val humidity: Double,
     val moon_phase: Double,
